pyproject.toml
[project]
name = "agentic-demo"
version = "0.1.0"
description = "Agentic workflow with LangChain, LangGraph, and local GPU support"
requires-python = ">=3.9"
authors = [
    { name = "Author", email = "author@example.com" }
]

dependencies = [
    "langchain>=0.1.13",
    "langgraph>=0.0.29",
    "langchain-community>=0.0.28",
    "langchain-core>=0.1.33",
    "requests>=2.31.0",
    "torch>=2.1.2",
    "torchvision>=0.16.2",
    "torchaudio>=2.1.2",
    "ollama>=0.1.16",
    "python-dotenv>=1.0.0",
    "chainlit>=2.3.0",
    "langchain-ollama>=0.3.10",
    "langchain-openai>=0.0.5",
    "openai>=1.0.0",
    "plotly>=6.5.0",
    "beautifulsoup4>=4.14.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "models", "tools"]

[tool.uv]
required-version = ">=0.9.5"
